+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # id3 BioSeal SDK Samples Changelog
 
-## Version 1.11.1.1 (2023-03-13)
+## Version 1.11.2 (2023-03-17)
+- Added c++ Linux and Windows sample.
+- Simplify dotnet sample.
+
+## Version 1.11.1.1 (2023-03-03)
 - Added new informations to display for id3 BioSeal SDK 1.11.1
 
 ## Version 1.6.3.1 (2023-01-12)